requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.1
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
anthropic==0.39.0
feedparser==6.0.11
apscheduler==3.10.4